refactor - Rename package name elements to domain
