+++ src/main/java/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar.elements;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
